adjust ToRomaji
