--- src/StringExRomajiToKatakana.cs
@@ -1,11 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace MyNihongo.KanaConverter;
-
-public static class StringExRomajiToKatakana
-{
-	
-
-
-	
-}
